apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Timing Agent API - 聚焦选股与择时的简化多智能体系统"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "akshare>=1.9.99",
  "litellm>=1.40.0",
  "python-dotenv>=1.0.0",
  "numpy>=1.26.0",
  "pandas>=2.1.0",
]

[scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
build = "python -m compileall app && mkdir -p dist && echo > dist/.gitkeep"
lint = "python -m compileall app"
